+++ app/models/guild_invite.rb
@@ -1,0 +1,5 @@
+# Represents invite to guild and requests for guild membership
+class GuildInvite < ApplicationRecord
+  belongs_to :guild
+  belongs_to :character
+end